transformation_pipeline/requirements.txt
colorama==0.4.4
numpy==1.22.2
pandas==1.4.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
scipy==1.8.0
six==1.16.0
tqdm==4.62.3
bvh==0.3